+++ extensions/posthog/helpers/useUrl.tsx
@@ -1,6 +1,5 @@
-import { getPreferenceValues } from "@raycast/api";
+import { AuthenticatedPostHogAccount } from "./posthog-auth";
 
-export function useUrl(path: string) {
-  const { dataRegionURL } = getPreferenceValues();
-  return `${dataRegionURL}/${path}`;
+export function useUrl(path: string, account: AuthenticatedPostHogAccount | null) {
+  return account ? `${account.baseUrl}/${path}` : "";
 }